[design] Corrected: - the Harmful Data Exposure of ClusterIO fields - the Data Clumps from several methods in appState.cpp - several Feature Envy issues by creating methods in the envied types (BestMatch, MatchParams) or in newly derived types (Centroid) - Underused Interfaces
These were issues reported by AI Reviewer.

Furthermore:
- ensured AI Reviewer will handle appropriately the variables defined previously as `auto`, the methods referred in lambda-s / calling chains / template functions, or those methods whose signatures contain std::string / std::unique_ptr / std::shared_ptr
- fixed several more warnings above level 3

+++ src/extBoxBlur.cpp
@@ -133,8 +133,8 @@ class ExtBoxBlur::Impl {
 					const double *dataItBegin = temp.ptr<double>(row);
 					const double *frontEdge = dataItBegin + kernelRadiusP1;
 					const double firstPixel = *dataItBegin,
-								lastPixel = dataItBegin[dataColsM1];
-					double prevFrontEdgePixel = dataItBegin[kernelRadius], frontEdgePixel,
+								lastPixel = dataItBegin[(size_t)dataColsM1];
+					double prevFrontEdgePixel = dataItBegin[(size_t)kernelRadius], frontEdgePixel,
 							prevSum = *resultIt =
 								w * (firstPixel + prevFrontEdgePixel) +
 								boxHeight * (firstPixel * kernelRadius + accumulate(dataItBegin + 1, dataItBegin + kernelRadius, 0.));
